+++ dipu/scripts/autogen_diopi_wrapper/diopi_wrapper_template.py
@@ -50,12 +50,12 @@ diopi_wrapper_file_template_content = """// autogened file
 #include "csrc_dipu/aten/ops/DIPUCopy.hpp"
 #include "csrc_dipu/aten/ops/NodispatchUtils.hpp"
 #include "csrc_dipu/aten/ops/OpUtils.hpp"
+#include "csrc_dipu/aten/ops/DIPUOpInferrer.h"
 #include "csrc_dipu/base/basedef.h"
 #include "csrc_dipu/diopirt/diopirt_impl.h"
 #include "csrc_dipu/profiler/profiler.h"
 #include "csrc_dipu/runtime/core/DIPUGeneratorImpl.h"
 #include "csrc_dipu/runtime/core/DIPUStream.h"
-#include "csrc_dipu/runtime/core/DIPUOpInferrer.h"
 
 #include "CustomFallbackFunctions.hpp"
